Use "-pthread" for GoogleTest builds in Emscripten
Add the "-pthread" flag when compiling and linking the GoogleTest static
libraries in Emscripten builds.

Meanwhile Pthreads aren't actually used by GoogleTest itself (due to the
other parameter that we specify), it's still necessary because we use
"-pthread" when compiling other libraries and source files, and
Emscripten requires this flag to be used consistently everywhere.

This is a preparatory fix for introducing Emscripten builds of unit
tests, as tracked by #177.