Substitute ArrayBuffer in cross-extension messages (#226)
Replace ArrayBuffer's with byte Array's when sending cross-extension
messages.

The background for this change is that Chrome Extension APIs don't
support sending ArrayBuffer's inside messages from one extension/app to
another. Currently the code in this project never attempts to do that,
but the refactoring proposed in #222 will start putting ArrayBuffer's
into the sent messages. Therefore the current commit anticipates this
upcoming change and adds an automatic conversion that will prevent data
loss in cross-extension messages.