Introduce "make run_test" for running tests
Rename the existing "run" target into "run_test", in order to avoid
clashing with the Native Client's definition of the "run" target.

Previously we were intentionally using this clash in order to reduce
amount of boilerplate needed for running tests, but that came with the
price of doing some nontrivial hacks (like changing the current
directory CURDIR back and forth). This commit gets rids of these hacks
and makes tests' makefiles slightly simpler.

This is a preparatory fix for introducing Emscripten builds of unit
tests, as tracked by #177.