Switch C++ sources format to Chromium code style
Change the format-code.sh script to tell clang-format to use the
"Chromium" style instead of the "Google" one. Also reformat all of the
existing code according to this new style (except the third-party code).

Given that the main product of this repository is the Smart Card
Connector app which is one of Chrome OS components (even if optional and
separately distributed), and given that all of this repository's
maintainers are also Chromium developers, it makes sense to switch to
the "Chromium" style guide.

Subjectively speaking, a big benefit of the "Chromium" style is that it
doesn't collapse short "if" statements into a single line - a two-line
format is, arguably, easier to read.