Fix goog.async.Deferred not available in JS code
This commit adds the
//third_party/closure-library/src/third_party/closure directory into the
list of source directories for JavaScript builds performed by Closure
Compiler.

This fixes the issue with inability to use the goog.async.Deferred
class, since it's defined in a file that was under this third_party
directory of the Closure Library.

This is a preparatory CL for introducing Emscripten/WebAssembly support
into the Example C++ App (tracked by #220), since the proposed way of
loading the Emscripten module will involve goog.async.Deferred.