Implement Emscripten entry point in C++ Example
Add C++ and JavaScript code into the C++ Example App
(//example_cpp_smart_card_client_app) that provide basic functionality
for loading the Emscripten module and interacting between the C++ and
the JavaScript code. This contributes to the WebAssembly migration
effort, as tracked by #220.

Note that this is not a functional build yet, since neither the other
C++ code is fully migrated to Emscripten, nor the JavaScript code for
Emscripten is wired up with the App's background.js script. These will
be addressed in follow-up CLs.