ifdef NaCl code in scard_structs_serialization (#250)
Surround the Native Client specific code with #ifdef's in
scard_structs_serialization.{h|cc}, so that it's only compiled in Native
Client builds. This way, these files can be used both by code that is
fully migrated to Emscripten/WebAssembly and the code that is still
using Native Client specific primitives.

This is a temporary step in order to accomplish the migration of the
//example_cpp_smart_card_client_app App (tracked by #220) earlier,
before the Smart Card Connector App gets migrated as well. Later,
once everything is migrated, all these #ifdef's together with the
code inside them will be removed.