Implement message channel for Emscripten JS loader (#254)
Extend the JavaScript "EmscriptenModule" class to have a bidirectional
message pipe with the executable. This message pipe conforms to the
Closure Library's goog.messaging.MessageChannel interface. Messages to
the module are sent by calling the "postMessage()" method on the
GoogleSmartCardModule class exposed by it via Embind. Messages in the
reverse direction are received via the callback passed to the class'
constructor.

This contributes to the WebAssembly migration effort, as tracked
by #220.