Post Emscripten JS messages only from main thread
This changes the GlobalContextImplEmscripten::PostMessageToJs()
implementation to forward the operation to the main Emscripten thread.
This fixes the inability to use this method from background threads,
since they're not allowed to make JavaScript operations like calling
back to the GoogleSmartCard.EmscriptenModule JavaScript class.

This contributes to the WebAssembly migration effort, as tracked
by #220.