Introduce top-level Makefile (#261)
This commit adds a new makefile, located in the project's root
directory. This makefile can be used to build the whole project in the
specified configuration, namely the TOOLCHAIN (can be "pnacl" [default]
or "emscripten") and the CONFIG (can be "Release" [default] or "Debug").

The main benefit is that the makefile syntax is more suitable than shell
scripts for describing rules like "target foo should always be compiled,
but target bar only under toolchain X" and similar more complex ones.
The makefile also naturally describes dependencies and prerequisites, as
opposed to hardcoding the "magically working" order of targets in shell
scripts.

Currently, only a few targets can be built in the TOOLCHAIN=emscripten
configuration. Follow-up changes will migrate the remaining targets, so
that everything can be compiled in all configurations.

Potentially, the //make-all.sh, //clean-all.sh, //test-all.sh scripts
will be rewritten to use this new makefile. This will make those scripts
simpler and keep the information about targets and their dependencies in
a single place.

This commit contributes to the WebAssembly migration effort, tracked
by #177.