Introduce ExecutableModule JS interface (#260)
Add the "ExecutableModule" JS abstract class that represents an
executable module; the interface allows to load it and send/receive
messages. Migrate the NaclModule class and all consumers to this
interface.

The core idea is that most of the JS code doesn't need to know whether
the module is based on Emscripten/WebAssembly or Native Client. All such
code should use the ExecutableModule interface.

A follow-up will migrate the EmscriptenModule class to this interface as
well.

This commit contributes to the WebAssembly migration effort, tracked
by #220.