Put built JS files into dir depending on toolchain (#274)
Make compiled JavaScript files be stored in different locations
depending on the TOOLCHAIN environment variable.

This fixes the problem that JS files can stay non-recompiled after
switching the TOOLCHAIN variable. E.g., in the following scenario
resulting JS files would still refer to emscripten if built without this
commit:

TOOLCHAIN=emscripten make && TOOLCHAIN=pnacl make

The underlying reason is that when both the input files are unchanged
and the resulting file paths don't depend on the toolchain, Make
doesn't see any reason to trigger recompilation. We don't have
prerequisites that depend on the compilation flags, and this would
be tricky to implement.

The solution in this commit makes sure that JS files are put into
different locations for pnacl and emscripten, which bypasses the problem
with missing recompilation. The only downside is a bit of uglyness for
the JS files that don't depend on the toolchain, but it's only a problem
for intermediate build files.

This commit contributes to the WebAssembly migration effort tracked
by #177.