Migrate libusb port to Emscripten
This commit allows to build our port of the libusb library under
Emscripten.

For the background, our libusb implementation is header-compatible with
the third-party libusb library, but instead of talking to drivers and
OS-level components it redirects all requests to the chrome.usb
Extensions API. Before this commit, our port could only be compiled
under Native Client.

This commit contributes to the effort of migrating the Smart Card
Connector app to Emscripten, as tracked by #233.