Make clean-all.sh preserve env by default
Change the //clean-all.sh script to keep the //env intact by default.
This means that the SDKs downloaded by the //env/initialize.sh script
(Emscripten and Native Client SDK) are preserved.

This makes the clean-all.sh script more convenient to use, since
developers usually need to delete intermediate build artifacts but not
SDKs, and reinitializing SDKs is a time-consuming step.

The "-f" option is also added that reverts to the old behavior, cleaning
both build artifacts and the env.