Rewrite make-all.sh to use the top-level Makefile
This commit refactors the //make-all.sh script to delegate the actual
build functionality to the //Makefile script.

Besides simplifying the script and removing duplicated information about
directories to be built, this also adds support of the
Emscripten/WebAssembly builds into make-all.sh.

This commit contributes to the Emscripten/WebAssembly migration effort,
as tracked by #177.