Add conversion of Value to/from nullptr
Implement a special case in the to/from Value conversions that converts
a null value into nullptr and vice versa.

This is a "fake" conversion in the sense that it's only used to denote
the type being null - there's no other payload involved. This special
case will be used in follow-up commits in order to support functions
whose input or output argument must always be null (such functions exist
in the PC/SC-Lite API).

This commit contributes to the Emscripten/WebAssembly migration effort
tracked by #185.