Support optional in RemoteCallArgumentsExtractor
Extend the RemoteCallArgumentsExtractor class to support extracting
optional arguments: for such arguments, the Value can be equal to null,
which denotes that the optional argument is unset.

This capability will be used in follow-up commits in order to support
receiving remote calls for functions with optional arguments.

This commit contributes to the Emscripten/WebAssembly migration effort,
as tracked by #233.