Add conversion of Value to/from nullptr (#290)
Implement a special case in the to/from Value conversions that converts
a null value into nullptr and vice versa.

This is a "fake" conversion in the sense that it's only used to denote
the type being null - there's no other payload involved. This special
case will be used in follow-up commits in order to support functions
whose input or output argument must always be null (such functions exist
in the PC/SC-Lite API). Previously, the Native Client's pp::Var::Null
type has been used for such arguments, but this commit adds a
toolchain-independent alternative that can be used in WebAssembly
builds.

This commit contributes to the Emscripten/WebAssembly migration effort
tracked by #185.