Build PCSC server_clients_management in Emscripten
This commit allows building the PC/SC-Lite server_clients_management
library in the Emscripten mode. Previously it was built only in for
Native Client builds, but due to the landed series of commits it can be
compiled using any of the two toolchains.

This commit contributes to the Emscripten/WebAssembly migration effort
tracked by #233.