Support Emscripten in Connector's background JS (#299)
Add support of Emscripten modules into the Smart Card Connector's
background page JavaScript code. With this commit, the JS behavior will
depend on the TOOLCHAIN constant (which, in turn, is filled based on the
environment variable), loading either the Emscripten or the Native
Client module.

This commit contributes to the Emscripten/WebAssembly migration effort,
as tracked by #233.