Build Connector app in Emscripten (#300)
This commit allows building the while Smart Card Connector app in the
Emscripten mode. Previously it was built only for Native Client builds,
but due to the landed series of commits it can be compiled using any
of the two toolchains now.

Note that the Emscripten build isn't fully functional yet - this will be
addressed by follow-up commits.

This commit contributes to the Emscripten/WebAssembly migration effort
tracked by #233.