Enable C++ exceptions in Emscripten builds (#301)
Enable the "DISABLE_EXCEPTION_CATCHING=0" flag when compiling C++ files
under Emscripten, so that the C++ exception support is enabled.

This flag was disabled by default since it affects performance and
binary size, but we're enabling this, as smart card middleware apps
reportedly rely on C++ exceptions. We don't set the flag only for
example_cpp_smart_card_client_app, since the flag must be used
consistently, which includes the common libraries that are linked into
it.

This commit contributes to the Emscripten/WebAssembly migration effort
tracked by #220.