Fix shutdown bug in PcscLiteServerClientsManager
Fix the crash that's happening in case
PcscLiteServerClientsManager::Detach() was used, that can happen during
the app shutdown.

The bug is that the destructor will be calling Detach() for the second
time, meanwhile that method is asserting against that. The fix is to
replace the assert with an early exit.

This commit contributes to the Emscripten/WebAssembly migration effort,
tracked by #233.