Omit "NaCl module" log prefix in Emscripten builds
This commit cleans up the log messages from the Emscripten modules by
removing the "NaCl module" prefix from them. This prefix will only be
printed in Native Client builds now.

For the background, the reason why this prefix is added under Native
Client is that the NaCl logs are redirected to Chrome's own stderr logs.
Therefore, to simplify reading the logs, we added this distinguished
prefix in NaCl builds. In Emscripten, on the contrary, such prefixes are
useless as Chrome doesn't embed Emscripten logs into its own logs.

This commit contributes to the Emscripten/WebAssembly migration effort,
as tracked by #185.