Fix shutdown bug in PcscLiteServerClientsManager (#303)
Fix the crash that's happening in case
PcscLiteServerClientsManager::Detach() was used, with the crash
potentially happenning during the app shutdown.

The bug is that the destructor will be calling Detach() for the second
time, meanwhile that method is asserting against being called multiple
times. The fix is to replace the assert with an early exit.

This commit contributes to the Emscripten/WebAssembly migration effort,
tracked by #233.