Cherry-pick PC/SC-Lite bugfix for thread function (#306)
Cherry-pick the following commit from the upstream PC/SC-Lite
repository:

  199f8455745654d3856251a949ad630c4d1a3c13
  Fix compiler warning

  hotplug_libusb.c: In function ‘HPSearchHotPluggables’:
  hotplug_libusb.c:498:4: warning: cast between incompatible function
  types from ‘void (*)(int *)’ to ‘void * (*)(void *)’
  [-Wcast-function-type]
    498 |    (PCSCLITE_THREAD_FUNCTION( )) HPEstablishUSBNotifications, pipefd);
        |    ^

This cherry-picked fix is necessary for the WebAssembly build of
PC/SC-Lite to work, since it fixes the wrong function signature that's
causing stack corruption when run under Emscripten. (For some reason,
this bug wasn't causing any issues when run as a native binary or under
Native Client.)

The cherry-picked patch file can be deleted once the upstream PC/SC-Lite
repository makes a new release that includes this commit.