Explicitly cast unowned release() to void (#312)
Add explicit casts to void in places that call
std::unique_ptr::release() without storing its result anywhere. This
clearly shows the programmer's intention, as opposed to potential places
where this could be done by mistake. This is a pure refactoring change,
in order to improve readability and to silence the clang-tidy
diagnostic.

This was found by the "bugprone-unused-return-value" clang-tidy
diagnostic.