Don't pass Emscripten linker flags to compiler (#325)
This commit attempts to avoid passing useless flags to the Emscripten
compiler and linker (emcc), by maintaining compilation and linking flags
separately instead of putting all of them together. This makes the build
logs slightly more readable (as the number of command-line flags becomes
less on average) and also avoids nonsensical no-op flags (like
EXPORT_NAME pointing to the name of a static library).

The main information about which Emscripten flags are needed on which
stage is present in this file:
https://github.com/emscripten-core/emscripten/blob/master/src/settings.js

This commit is expected to bring no functional differences on the
resulting build program. The tracking issue is #177.