Rename nacl_module_dependency_in_app_common.mk (#329)
Rename this helper makefile to reflect the fact that this file is
toolchain-agnostic, i.e., can be used with either Emscripten or Native
Client builds. Also rename the rules inside this file accordingly, and
drop unneeded assertions on NaCl.

This is a non-functional change. This contributes to the WebAssembly
migration effort tracked by #177.