Use g3 instead of g4 with Emscripten
Replace the "-g4" compiler/linker flag with "-g3" when compiling via
Emscripten. Apparently, "g3" started to produce more debug information
that "g4" in recent Emscripten versions, in particular for the recently
added DWARF debug format:
https://github.com/emscripten-core/emscripten/issues/13534

Note that the debugging source maps are only produced at "g4", but we
never supported them properly in this project (in particular, due to
DevTools not being able to load them in Extensions/Apps).

This contributes to the WebAssembly/Emscripten migration effort, as
tracked by #177.