Log function call indices
Print a function call index (a shared monotonically incrementing
counter) when logging function call tracer.

This should simplify investigating log messages collected in the field,
since it'll be possible to correlate the "called..." and the "returning"
log messages, therefore knowing which input parameters were passed to
the completed call and being able to distinguish between multiple calls
with the same parameters.