Remove relative time from JS logs (#339)
Stop printing relative times (like "[ 0.007s]") in the JavaScript logs.
These times aren't really necessary now that we're printing absolute
times since #108.

The only remaining benefit of relative times was that they allowed
to distinguish logs from multiple instances in case they work
concurrently and their logs are squashed together. However, in
practice this was only relevant in a specific scenario (reading
Chrome OS system logs from a device that is locked and has the
NaCl-based smart card applications installed both in-session
and on the Login Screen). Assuming this a rare scenario that will
anyway disappear with the NaCl deprecation, it seems fine to
reduce the log clutter by removing the relative times.

This change contributes to the logging improvements tracked
by #146.