Detect NaCl crash loops before Chrome throttles us (#353)
Improve the crash loop detector for the case when the crash happens in
the C/C++ code compiled under Native Client. Before this commit, the
crash loop detection would be triggered too late, after Chrome Native
Client system already began the throttling. This means that instead of
useful logs that contain information before the actual crash, the only
information would be like this:

>   Failed to load NaCl module with the following message: NaCl module
>   load failed: ServiceRuntime: failed to start

With the following information in Chrome logs:

>  NaCl process launch failed: Process creation was throttled due to
>  excessive crashes

This commit avoids this situation by tightening the crash loop detection
constants, so that the detector is triggered earlier and therefore
avoids this dead loop and preserves valuable logs.

P.S. Meanwhile this commit is applied to both NaCl and WebAssembly
builds, it's presumably OK, since all we do is tightening the crash loop
detection constants. If we discover later that WebAssembly contains a
similar, but even more aggressive throttling, we can tighten the
constants further; otherwise they don't harm.