Workaround for chrome.usb.getConfigurations issue
Change the USB wrapper (used in the webport of libusb, which is
transitively used for the CCID driver and therefore the Smart Card
Connector app) to stop asserting that chrome.usb.getConfigurations() can
only return a single active configuration. Instead, just proceed with
the first configuration among those that are marked as active.

Meanwhile multiple active configurations is most likely a bug in the
chrome.usb API (https://crbug.com/1218141), we need this workaround
until that issue is fixed in Chrome, since currently the Connector app
ends up in a crash loop when such device is attached.