Update Emscripten to 2.0.24 (#359)
Switch the version hardcoded in the env initialization scripts to use
the currently up-to-date version of Emscripten. It brings various
improvements that the Emscripten team made recently, and at the same
time doesn't regress on our use cases (we couldn't update earlier due to
a few blockers in the previous releases, like Clang crashes).