Enable *.js reformatting in format-code.sh (#365)
Change the format-code.sh developer script to also automatically reformat
any change JavaScript file.

In the past, all JS code written in this repository was formatted
manually. The plan is to switch to the clang-based automatic formatting,
and the first step on this route is to apply it to every changed file.

This contributes to the refactoring effort tracked by #264.