Add missing @const
Refactor JavaScript code to specify "@const" for all class attributes
that are never reassigned. Also promote the "jsdocMissingConst" Closure
Compiler diagnostic, which complained about private properties that can
be marked as @const, to be an error.

This is intended to be a non-function change, contributing to the
refactoring effort tracked by #264.