Migrate DeferredProcessor to ES6 classes
Stop using the old-style Closure Compiler classes and switch to modern
ES6 classes. Code was *not* reformatted, in order to make the diff
readable; later there'll be a separate commit that'll apply clang-tidy.

A brief overview of the changes:
* Instead of declaring the class by marking a function with
  "@constructor", the "class" ES6 keyword is used.
* Instead of using goog.inherits() for declaring the inheritance, the
  "extends" ES6 keyword is used.
* Static properties are first defined as file-global constants and only
  then copied into the class; this allows using them in the class code,
  since the old-style "this.someConstant" access doesn't work anymore.
  (A side note: ES6 class static properties aren't supported by Closure
  Compiler yet.)
* Calls to the parent class' methods are done via super() instead of the
  Closure Compiler's base() method.

This is expected to be a non-functional change. It contributes to the
refactoring effort tracked by #264.