Add built-time PACKAGING environment variable
This variable can be specified in the command line when running "make"
and specify the needed mode. By default, the "app" mode is used. The
value specified in the command line is exposed to the JavaScript code as
the GoogleSmartCard.Packaging.MODE constant.

This contributes to the app-to-extension migration effort, in particular
task #379.