Avoid exceptions on missing USB API
Improve the ChromeUsbBackend JavaScript code to avoid throwing
exceptions when the USB API is actually unavailable.

This should never happen in production, but gracefully handling the
errors helps in the early stage of the app-to-extension migration of the
Smart Card Connector. Also an extra safety won't harm anyway.

This contributes to the app-to-extension migration effort, in particular
task #383.