app/extension conditions in manifest.json.template (#385)
Add the support of conditionals into the Connector's
manifest.json.template, allowing to switch between extension- and
app-specific blocks in this file depending on the PACKAGING build
variable.

Do this by introducing a simple "${if ...} ... ${endif}" syntax into
create_app_manifest.py, with the regexp-based implementation. The
implementation is very basic and, e.g., doesn't support nested
conditions.

This contributes to the app-to-extension migration effort, in particular
task #379.