Fix log collection in Connector's extension mode
This fixes the log collection and exporting functionality in the
"extension" packaging mode of the Smart Card Connector. The goal is to
collect the logs from all windows and pages in the background page, and
also to allow exporting them when the user clicks "Export logs" in the
UI.

Implementation-wise, this is achieved by switching to a different
mechanism for communicating between the background page and other
windows. The old app-specific mechanism worked like this: the background
page holds a log buffer that it injects into every opened window by
setting a global property on it. The new mechanism used in this commit
is the inverse of this: every window calls
chrome.runtime.getBackgroundPage() and grabs the log buffer.

Some additional complexity in this change is caused by the fact that
getBackgroundPage() is asynchronous, so that we need to preserve and
then move all logs we've accumulated till it completes. Also we need to
be careful to register the log buffer in the window's logger, which
means using the window's Closure Library instance.

This contributes to the app-to-extension migration effort, in particular
task #390.