Fix ArrayBuffer bug in single-message-based comm
Fix the issue with incorrect serialization of byte fields when using
single-message-based communication channels. The root cause is that
we did internal migration in the Smart Card Connector app to start using
ArrayBuffers for representing such fields, however due to the
cross-extension messages not supporting them we needed to transform this
into integer arrays before sending this. While #226 did this for
port-based communications, the single-message-based communication left
broken, resulting in some fields, like "atr", being received as "{}"
instead of the array of integer-bytes.

This fixes #393.