Don't call getBackgroundPage in non-app/extensions (#396)
Skip calling chrome.runtime.getBackgroundPage() when setting up logging
(originally added in #391), so that this code doesn't throw an exception
when this method is unavailable. This should never happen in production
currently, but might happen if the code is running outside an app/extension.

This contributes to the app-to-extension migration effort, in particular
task #390.