Add helper for checking string being extension ID
Add a small helper for checking whether a string is (purely based on its
format) an extension ID.

The plan is to use this helper in follow-ups in order to distinguish
between extension IDs and generic web origins. Specifically, we're
adding the support of receiving smart card commands from web pages,
which means we'll store two kinds of permissions granted by the user:
ones based on Extension IDs and ones based on web origins; when loading
these persisted permissions, we'll need to determine the kind of every
stored item.

This is a preparatory commit for the effort tracked by #377: supporting
PC/SC requests from web pages.